Extract Kore.Builtin.List.List
